move API contact and license to gateway
