+++ api-gateway/src/main/java/gov/va/vro/propmodel/OpenApiProperties.java
@@ -1,0 +1,18 @@
+package gov.va.vro.propmodel;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+@Getter
+@Setter
+@ConfigurationProperties(prefix = "vro.openapi")
+public class OpenApiProperties {
+  private final Info info = new Info();
+
+  private List<Server> servers = new ArrayList<Server>(Arrays.asList(new Server()));
+}
