+++ api-gateway/src/main/java/gov/va/vro/propmodel/Server.java
@@ -1,0 +1,11 @@
+package gov.va.vro.propmodel;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Server {
+  private String description = "Local Server";
+  private String url = "/";
+}
