+++ app/src/main/java/gov/va/vro/config/propmodel/OpenApiProperties.java
@@ -1,7 +1,9 @@
 package gov.va.vro.config.propmodel;
 
+import gov.va.vro.config.propmodel.Server;
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -9,8 +11,7 @@ import java.util.List;
 
 @Getter
 @Setter
-public class OpenApi {
-  private final Info info = new Info();
-
+@ConfigurationProperties(prefix = "vro.openapi")
+public class OpenApiProperties {
   private List<Server> servers = new ArrayList<Server>(Arrays.asList(new Server()));
 }
